+++ core/entity/update.js
@@ -1,0 +1,11 @@
+
+module.exports = function(updates, next) {
+
+  for( var prop in updates ) {
+    this[prop] = updates[prop];
+  }
+  
+  this.save( err => next(err) );
+
+}
+