+++ core/service/script.js
@@ -1,14 +1,15 @@
-var Script = require(process.env.BASE_PATH + "/entity/script").Entity;
-var Task = require(process.env.BASE_PATH + "/entity/task").Entity;
-var logger = require('../lib/logger')('eye:supervisor:service:script');
-var S3Storage = require('../lib/store').S3;
-var LocalStorage = require('../lib/store').Local;
 var async = require('async');
 var md5 = require('md5');
 var fs = require('fs');
 var config = require('config');
-var elastic = require('../lib/elastic');
 var path = require('path');
+var extend = require('util')._extend;
+var Script = require("../entity/script").Entity;
+var Task = require("../entity/task").Entity;
+var logger = require('../lib/logger')('eye:supervisor:service:script');
+var S3Storage = require('../lib/store').S3;
+var LocalStorage = require('../lib/store').Local;
+var elastic = require('../lib/elastic');
 
 var storageMedia = (function getStorageMedia(){
   return config.get("storage").driver == 'local' ?
@@ -130,34 +131,33 @@ var Service = {
     var script = input.script;
     var file = input.file;
 
-    updateScriptFile(file,script,
-      function(error, data){
-        var updates = {
-          'description': input.description,
-          'keyname': data.keyname,
-          'md5': data.md5,
-          'name': file.name,
-          'mimetype': file.mimetype,
-          'size': file.size,
-          'extension': file.extension,
-          'public': input.public
-        };
+    updateScriptFile(file, script, (error, data) => {
+      var updates = extend(input,{
+        'keyname': data.keyname,
+        'md5': data.md5,
+        'filename': file.name,
+        'mimetype': file.mimetype,
+        'size': file.size,
+        'extension': file.extension,
+        'last_update': new Date()
+      });
 
-        logger.log('updating script data');
-        script.update(updates, function(error){
-          if(error) return next(error);
-          registerScriptCRUDOperation(input.customer.name,{
+      logger.log('updating script');
+      script.update(updates,error => {
+        if(error) return next(error);
+        registerScriptCRUDOperation(
+          input.customer.name, {
             'name':script.name,
             'customer_name':input.customer.name,
             'user_id':input.user.id,
             'user_email':input.user.email,
             'operation':'update'
-          });
-          logger.log('script update completed');
-          next(null,script);
-        });
-      }
-    );
+          }
+        );
+        logger.log('script updated');
+        next(null,script);
+      });
+    });
   },
   remove : function(input,next)
   {