+++ core/entity/monitor/index.js
@@ -7,23 +7,20 @@ var BaseSchema = require('./schema');
 var Template = require('./template').Entity;
 var Resource = require('../resource').Entity;
 var logger = require('../../lib/logger')('eye:entity:monitor');
-var _ = require('lodash');
 
 var properties = {
-  'host_id': { type: String, required: true },
-  'resource': { type: ObjectId, ref: 'Resource' },
-  'resource_id': { type: String },
-  'enable': { type: Boolean, 'default': true },
-  'creation_date': { type: Date, 'default': Date.now },
-  'last_update': { type: Date, 'default': Date.now },
-  'template': { type: ObjectId, ref: 'MonitorTemplate', 'default': null },
+  host_id: { type: String, required: true },
+  host: { type: ObjectId, ref: 'Host' },
+  resource: { type: ObjectId, ref: 'Resource' },
+  resource_id: { type: String },
+  enable: { type: Boolean, 'default': true },
+  creation_date: { type: Date, 'default': Date.now },
+  last_update: { type: Date, 'default': Date.now },
+  template: { type: ObjectId, ref: 'MonitorTemplate', 'default': null }
 }
 
 /** Extended Schema. Includes non template attributes **/
 var MonitorSchema = BaseSchema.EntitySchema.extend(properties);
-
-/** Exports all the properties **/
-//exports.properties = _.extend({},BaseSchema.properties,properties);
 
 /**
  * extends publishing method to include Entity specific definitions
@@ -33,7 +30,9 @@ MonitorSchema.methods.publish = function(options, next)
 {
   options = options || {};
   if( options.populate ){
-    Entity.populate(this, { path:'resource' }, function(error, monitor){
+    Entity.populate(this, {
+      path:'resource'
+    }, function(error, monitor){
       if(!monitor.resource) {
         logger.error('monitor.resource is null. could not populate');
         next(error);